+++ lib/VoicePatternData/src/VoicePatternData.h
@@ -6,10 +6,10 @@
 class VoicePatternData {
 public:
   
-  void at(uint32_t value, uint8_t position);
-  uint32_t at(uint8_t position);
+  void at(uint8_t value, uint8_t position);
+  uint8_t at(uint8_t position);
   void resize(uint8_t newLength);
 private:
-  std::vector<uint32_t> _voicePattern;
+  std::vector<uint8_t> _voicePattern;
 };
 #endif
